Aggiunto mpm test ma ancora da sistemare 3 errori

+++ ViewProject/ViewProject/Persistence/MainPersistenceManager.cs
@@ -219,6 +219,8 @@ namespace ViewProject.Persistence
 
         public bool DeleteAllenamenti(Utente utente)
         {
+            if (utente == null)
+                throw new ArgumentException();
             try
             {
                 SqlCommand delete = new SqlCommand("delete from ALLENAMENTI where username = '" + utente.Username + "' ;", Conn);
@@ -232,6 +234,8 @@ namespace ViewProject.Persistence
 
         public bool DeletePianoAllenamento(Utente utente)
         {
+            if (utente == null)
+                throw new ArgumentException();
             //grazie ad on delete cascade mi si eliminano anche tutti le tuple referenziate
             try
             {
@@ -246,6 +250,8 @@ namespace ViewProject.Persistence
 
         public bool DeleteUtente(Utente utente)
         {
+            if (utente == null)
+                throw new ArgumentException();
             try
             {
                 SqlCommand delete = new SqlCommand("delete from UTENTI where username='" + utente.Username + "';", Conn);
@@ -259,6 +265,8 @@ namespace ViewProject.Persistence
 
         public IEnumerable<Allenamento> LoadAllAllenamenti(Utente utente)
         {
+            if (utente == null)
+                throw new ArgumentException();
             List<Allenamento> allenamenti = new List<Allenamento>();
             try
             {
@@ -294,6 +302,8 @@ namespace ViewProject.Persistence
 
         public bool CheckUsername(string username)
         {
+            if (String.IsNullOrEmpty(username))
+                throw new ArgumentException();
             try
             {
                 SqlCommand select = new SqlCommand("delete from UTENTI where username='" + username + "';", Conn);
@@ -310,6 +320,8 @@ namespace ViewProject.Persistence
 
         public bool ThereIsAPianoAllenamento(Utente utente)
         {
+            if (utente == null)
+                throw new ArgumentException();
             try
             {
                 bool res;
@@ -331,6 +343,8 @@ namespace ViewProject.Persistence
 
         public PianoAllenamento LoadPianoAllenamento(Utente utente)
         {
+            if (utente == null)
+                throw new ArgumentException();
             PianoAllenamento pianoAllenamento;
             try
             {
@@ -406,6 +420,8 @@ namespace ViewProject.Persistence
 
         public Utente Autentica(string username, string password)
         {
+            if (String.IsNullOrEmpty(username) || String.IsNullOrEmpty(password))
+                throw new ArgumentException();
             try
             {
                 Utente utente;
@@ -463,6 +479,8 @@ namespace ViewProject.Persistence
 
         public bool SaveAllenamento(Utente utente, Allenamento allenamento)
         {
+            if (utente == null || allenamento == null)
+                throw new ArgumentException();
             try
             {
                 int allenamentoID = _IDBroker.generaAllenamentoID();
@@ -512,6 +530,8 @@ namespace ViewProject.Persistence
 
         public bool SavePianoAllenamento(Utente utente, PianoAllenamento pianoAllenamento)
         {
+            if (utente == null || pianoAllenamento == null)
+                throw new ArgumentException();
             if (ThereIsAPianoAllenamento(utente))
             {
                 DeletePianoAllenamento(utente);
@@ -624,6 +644,8 @@ namespace ViewProject.Persistence
         }
         public bool updateUtente(Utente utente)
         {
+            if (utente == null)
+                throw new ArgumentException();
             try
             {
                 //uso di SqlParameter per garantire sicurezza e evitare Sql Injection
@@ -684,6 +706,8 @@ namespace ViewProject.Persistence
 
         public bool SaveUtente(Utente utente, string password)
         {
+            if (utente == null || String.IsNullOrEmpty(password))
+                throw new ArgumentException();
             try
             {
                 //uso di SqlParameter per garantire sicurezza e evitare Sql Injection
@@ -736,6 +760,8 @@ namespace ViewProject.Persistence
 
         public bool SaveUtenteAutomatico(Utente utente, Risorsa risorsa, int numeroGiorniAllenamento, TipoAllenamento tipo)
         {
+            if (utente == null || numeroGiorniAllenamento<=0 || numeroGiorniAllenamento>7)
+                throw new ArgumentException();
             try
             {
                     SqlCommand insertUtenteAutomatico;
@@ -803,7 +829,7 @@ namespace ViewProject.Persistence
             }
             catch (SqlException)
             {
-                throw new Exception("ciao");
+                throw new Exception();
             }
         }
 
@@ -822,6 +848,8 @@ namespace ViewProject.Persistence
 
         public static TipoAllenamento? getTipoAllenamento(string tipoString)
         {
+            if (String.IsNullOrEmpty(tipoString))
+                throw new ArgumentException();
             TipoAllenamento? tipo = null;
             switch (tipoString.ToLower())
             {
@@ -842,6 +870,8 @@ namespace ViewProject.Persistence
 
         public static Sesso? getSesso(string sessoString)
         {
+            if (String.IsNullOrEmpty(sessoString))
+                throw new ArgumentException();
             Sesso? sesso = null;
             switch (sessoString.ToLower())
             {
@@ -857,6 +887,8 @@ namespace ViewProject.Persistence
 
         public static Risorsa? getRisorsa(string risorsaString)
         {
+            if (String.IsNullOrEmpty(risorsaString))
+                throw new ArgumentException();
             Risorsa? risorsa = null;
             switch (risorsaString.ToLower())
             {
@@ -878,6 +910,8 @@ namespace ViewProject.Persistence
 
         public Esercizio GetEsercizioByName(string nome)
         {
+            if (String.IsNullOrEmpty(nome))
+                throw new ArgumentException();
             foreach (Esercizio esercizio in _esercizi)
             {
                 if (esercizio.Nome.ToLower().Equals(nome.ToLower()))
@@ -888,6 +922,8 @@ namespace ViewProject.Persistence
 
         public static FasciaMuscolare? getFasciaMuscolare(string fasciaMuscolareString)
         {
+            if (String.IsNullOrEmpty(fasciaMuscolareString))
+                throw new ArgumentException();
             FasciaMuscolare? fasciaMuscolare = null;
             switch (fasciaMuscolareString.ToLower())
             {
